+++ src/renderer/src/components/tab-group/useTabDragSplit.ts
@@ -1,6 +1,3 @@
-/* oxlint-disable max-lines -- Why: the drag-split hook co-locates drop-zone
- * resolution, same-group reordering, and cross-group handoff so state
- * transitions stay readable in one place. */
 import { useCallback, useRef, useState, type RefObject } from 'react'
 import {
   closestCenter,
@@ -16,38 +13,26 @@ import {
 } from '@dnd-kit/core'
 import type { TabGroup } from '../../../../shared/tab-types'
 import { useAppStore } from '../../store'
-import { mirrorWebRuntimeTabMove } from '../tab-bar/web-runtime-tab-move-mirror'
+import { useHoveredTabInsertion, type HoveredTabInsertion } from './tab-insertion'
 import {
-  resolveTabInsertion,
-  useHoveredTabInsertion,
-  type HoveredTabInsertion
-} from './tab-insertion'
-import { acquireWebviewsDragPassthrough } from '../browser-pane/webview-registry'
-import {
-  applyDragPreviewTab,
   captureTabDragActivationSnapshot,
   restoreSourceGroupActiveTabAfterCrossGroupDrop,
   restoreTabDragActivationSnapshot,
   type TabDragActivationSnapshot
 } from './tab-drag-preview-activation'
-import { resolveDragPreviewTabId, resolveSourceGroupRestoreOnDrop } from './tab-drag-preview-target'
 import { getDragPointer } from './tab-drag-pointer'
 import { TabDragPointerSensor } from './tab-drag-pointer-sensor'
 import {
   captureTabGroupPanelGeometrySnapshot,
-  resolveActivePaneColumnSplitTarget,
-  type ActivePaneColumnSplitTarget,
   type TabGroupPanelGeometrySnapshot
 } from './tab-group-panel-split-target'
-import {
-  canDropTabIntoPaneBody,
-  isPaneDropData,
-  isTabDragData,
-  type TabDragItemData,
-  type TabDropZone
-} from './tab-drag-data'
+import { canDropTabIntoPaneBody, isTabDragData, type TabDragItemData } from './tab-drag-data'
+import { useTabDragGestureLifecycle } from './tab-drag-gesture-lifecycle'
+import { useTabDragHoverPreview, type HoveredTabDropTarget } from './tab-drag-hover-preview'
+import { commitTabDragDrop } from './tab-drag-drop-commit'
 
 export type { HoveredTabInsertion }
+export type { HoveredTabDropTarget }
 export {
   canDropTabIntoPaneBody,
   isPaneDropData,
@@ -60,12 +45,6 @@ export {
 // Why: tab activation waits for pointerup, so dnd-kit needs enough movement
 // tolerance to avoid treating ordinary click jitter as an intentional drag.
 export const TAB_DRAG_ACTIVATION_DISTANCE_PX = 12
-
-export type HoveredTabDropTarget = {
-  groupId: string
-  zone: TabDropZone
-  panelRect?: DOMRect
-}
 
 export function canDropTabForPaneColumnSplit(args: {
   activeDrag: TabDragItemData | null
@@ -123,15 +102,28 @@ export function useTabDragSplit({
   const reorderUnifiedTabs = useAppStore((state) => state.reorderUnifiedTabs)
   const dropUnifiedTab = useAppStore((state) => state.dropUnifiedTab)
   const [activeDrag, setActiveDrag] = useState<TabDragItemData | null>(null)
-  const [hoveredDropTarget, setHoveredDropTarget] = useState<HoveredTabDropTarget | null>(null)
-  const releaseWebviewDragPassthroughRef = useRef<(() => void) | null>(null)
   const preDragActivationSnapshotRef = useRef<TabDragActivationSnapshot | null>(null)
-  const lastPreviewRef = useRef<{ groupId: string; tabId: string | null } | null>(null)
-  const lastHoveredTabPreviewRef = useRef<{ groupId: string; tabId: string } | null>(null)
   const tabDragActiveRef = useRef(false)
   const dragGeometryRef = useRef<TabGroupPanelGeometrySnapshot | null>(null)
-  const releaseMissedEndFallbackRef = useRef<(() => void) | null>(null)
+  const clearDragStateRef = useRef<() => void>(() => {})
   const tabInsertion = useHoveredTabInsertion(isTabDragData, getDragPointer)
+  const {
+    acquireWebviewDragPassthrough,
+    installMissedEndFallback,
+    releaseMissedEndFallback,
+    releaseWebviewDragPassthrough,
+    setDragRootNode
+  } = useTabDragGestureLifecycle({ clearDragStateRef, tabDragActiveRef })
+  const {
+    clear: clearHoveredDropTarget,
+    handleDragUpdate,
+    hoveredDropTarget
+  } = useTabDragHoverPreview({
+    worktreeId,
+    preDragActivationSnapshotRef,
+    dragGeometryRef,
+    tabInsertion
+  })
 
   // Why: hidden worktrees stay mounted so their PTYs survive worktree
   // switches, but their DndContext should not activate drags. We use an
@@ -144,84 +136,21 @@ export function useTabDragSplit({
   })
   const sensors = useSensors(pointerSensor)
 
-  const releaseWebviewDragPassthrough = useCallback(() => {
-    releaseWebviewDragPassthroughRef.current?.()
-    releaseWebviewDragPassthroughRef.current = null
-  }, [])
-
-  const releaseMissedEndFallback = useCallback(() => {
-    releaseMissedEndFallbackRef.current?.()
-    releaseMissedEndFallbackRef.current = null
-  }, [])
-
-  const clearDragStateRef = useRef<() => void>(() => {})
-
-  const installMissedEndFallback = useCallback(() => {
-    releaseMissedEndFallback()
-
-    let cleanupTimer: number | null = null
-    const clearIfDndMissedEnd = (): void => {
-      if (cleanupTimer !== null) {
-        window.clearTimeout(cleanupTimer)
-      }
-      cleanupTimer = window.setTimeout(() => {
-        cleanupTimer = null
-        if (tabDragActiveRef.current) {
-          // Why: Electron/dnd-kit can occasionally miss drag end/cancel; a
-          // stuck drag ref makes all later tab clicks look like drag releases.
-          clearDragStateRef.current()
-        }
-      }, 0)
-    }
-
-    window.addEventListener('pointerup', clearIfDndMissedEnd)
-    window.addEventListener('pointercancel', clearIfDndMissedEnd)
-    window.addEventListener('blur', clearIfDndMissedEnd)
-    window.addEventListener('focus', clearIfDndMissedEnd)
-    releaseMissedEndFallbackRef.current = () => {
-      if (cleanupTimer !== null) {
-        window.clearTimeout(cleanupTimer)
-      }
-      window.removeEventListener('pointerup', clearIfDndMissedEnd)
-      window.removeEventListener('pointercancel', clearIfDndMissedEnd)
-      window.removeEventListener('blur', clearIfDndMissedEnd)
-      window.removeEventListener('focus', clearIfDndMissedEnd)
-    }
-  }, [releaseMissedEndFallback])
-
-  const acquireWebviewDragPassthrough = useCallback(() => {
-    // Why: dnd-kit tab drags are pointer-driven, so the native drag listeners
-    // in webview-registry never fire. Put webviews in passthrough explicitly.
-    releaseWebviewDragPassthrough()
-    releaseWebviewDragPassthroughRef.current = acquireWebviewsDragPassthrough()
-  }, [releaseWebviewDragPassthrough])
-
-  const setDragRootNode = useCallback(
-    (node: HTMLDivElement | null): void => {
-      if (node) {
-        return
-      }
-      // Why: this root owns the dnd-kit gesture that temporarily puts browser
-      // webviews in pointer passthrough and installs global fallback listeners,
-      // so root teardown must release both.
-      releaseWebviewDragPassthrough()
-      releaseMissedEndFallback()
-    },
-    [releaseMissedEndFallback, releaseWebviewDragPassthrough]
-  )
-
   const clearDragState = useCallback(() => {
     tabDragActiveRef.current = false
     releaseWebviewDragPassthrough()
     releaseMissedEndFallback()
     setActiveDrag(null)
-    setHoveredDropTarget(null)
+    clearHoveredDropTarget()
     tabInsertion.clear()
     preDragActivationSnapshotRef.current = null
-    lastPreviewRef.current = null
-    lastHoveredTabPreviewRef.current = null
     dragGeometryRef.current = null
-  }, [releaseMissedEndFallback, releaseWebviewDragPassthrough, tabInsertion])
+  }, [
+    clearHoveredDropTarget,
+    releaseMissedEndFallback,
+    releaseWebviewDragPassthrough,
+    tabInsertion
+  ])
   clearDragStateRef.current = clearDragState
 
   const restorePreDragActivation = useCallback(() => {
@@ -260,88 +189,6 @@ export function useTabDragSplit({
     [clearDragState, restorePreDragActivation, restoreSourceGroupAfterCrossGroupDrop]
   )
 
-  const updateDragPreviewActivation = useCallback(
-    (event: DragMoveEvent | DragOverEvent, activeData: TabDragItemData) => {
-      const snapshot = preDragActivationSnapshotRef.current
-      if (!snapshot) {
-        return
-      }
-
-      const overData = event.over?.data.current
-      if (isTabDragData(overData) && overData.unifiedTabId !== activeData.unifiedTabId) {
-        lastHoveredTabPreviewRef.current = {
-          groupId: overData.groupId,
-          tabId: overData.unifiedTabId
-        }
-      }
-
-      const preview = resolveDragPreviewTabId({
-        activeDrag: activeData,
-        overData,
-        preDragActiveTabIdByGroup: snapshot.activeTabIdByGroup,
-        lastHoveredTabPreview: lastHoveredTabPreviewRef.current
-      })
-      const lastPreview = lastPreviewRef.current
-      if (lastPreview?.groupId === preview.groupId && lastPreview.tabId === preview.tabId) {
-        return
-      }
-      lastPreviewRef.current = preview
-      applyDragPreviewTab({
-        worktreeId,
-        groupId: preview.groupId,
-        tabId: preview.tabId,
-        activeGroupId: preview.groupId
-      })
-    },
-    [worktreeId]
-  )
-
-  const updateHoveredDropTargetFromSplit = useCallback(
-    (splitTarget: ActivePaneColumnSplitTarget | null) => {
-      if (!splitTarget) {
-        setHoveredDropTarget((prev) => (prev === null ? prev : null))
-        return
-      }
-      setHoveredDropTarget((prev) => {
-        if (prev?.groupId === splitTarget.groupId && prev?.zone === splitTarget.zone) {
-          return prev
-        }
-        return {
-          groupId: splitTarget.groupId,
-          zone: splitTarget.zone,
-          panelRect: splitTarget.panelRect
-        }
-      })
-    },
-    []
-  )
-
-  const handleDragUpdate = useCallback(
-    (event: DragMoveEvent | DragOverEvent) => {
-      const activeData = event.active.data.current
-      if (isTabDragData(activeData) && activeData.worktreeId === worktreeId) {
-        updateDragPreviewActivation(event, activeData)
-      }
-
-      const state = useAppStore.getState()
-      const splitTarget = resolveActivePaneColumnSplitTarget({
-        event,
-        groupsByWorktree: state.groupsByWorktree,
-        layoutByWorktree: state.layoutByWorktree,
-        worktreeId,
-        getDragPointer,
-        geometry: dragGeometryRef.current
-      })
-      updateHoveredDropTargetFromSplit(splitTarget)
-      if (splitTarget) {
-        tabInsertion.clear()
-      } else {
-        tabInsertion.update(event)
-      }
-    },
-    [tabInsertion, updateDragPreviewActivation, updateHoveredDropTargetFromSplit, worktreeId]
-  )
-
   const onDragStart = useCallback(
     (event: DragStartEvent) => {
       const dragData = event.active.data.current
@@ -374,157 +221,16 @@ export function useTabDragSplit({
 
   const onDragEnd = useCallback(
     (event: DragEndEvent) => {
-      const activeData = event.active.data.current
-      const overData = event.over?.data.current
-      let shouldRestorePreDragActivation = true
-
-      if (!isTabDragData(activeData) || activeData.worktreeId !== worktreeId) {
-        finishDrag(true)
-        return
-      }
-
-      const state = useAppStore.getState()
-      const paneColumnSplit = resolveActivePaneColumnSplitTarget({
+      commitTabDragDrop({
         event,
-        groupsByWorktree: state.groupsByWorktree,
-        layoutByWorktree: state.layoutByWorktree,
         worktreeId,
-        getDragPointer,
-        geometry: dragGeometryRef.current
+        dragGeometryRef,
+        dropUnifiedTab,
+        reorderUnifiedTabs,
+        finishDrag
       })
-      if (paneColumnSplit) {
-        const moved = dropUnifiedTab(activeData.unifiedTabId, {
-          groupId: paneColumnSplit.groupId,
-          splitDirection: paneColumnSplit.zone
-        })
-        if (moved) {
-          shouldRestorePreDragActivation = false
-          mirrorWebRuntimeTabMove({
-            kind: 'split',
-            worktreeId,
-            tabId: activeData.unifiedTabId,
-            targetGroupId: paneColumnSplit.groupId,
-            splitDirection: paneColumnSplit.zone
-          })
-        }
-        finishDrag(
-          shouldRestorePreDragActivation,
-          resolveSourceGroupRestoreOnDrop(
-            activeData,
-            paneColumnSplit.groupId,
-            shouldRestorePreDragActivation
-          )
-        )
-        return
-      }
-
-      if (!event.over) {
-        finishDrag(true)
-        return
-      }
-
-      if (isTabDragData(overData)) {
-        if (activeData.unifiedTabId === overData.unifiedTabId) {
-          finishDrag(true)
-          return
-        }
-
-        const groups = state.groupsByWorktree[worktreeId] ?? []
-        const targetGroup = groups.find((group) => group.id === overData.groupId)
-        if (!targetGroup) {
-          finishDrag(true)
-          return
-        }
-
-        // Why: dnd-kit's `over` is the hovered tab, but the drop's true
-        // insertion point depends on which side of that tab the cursor sits.
-        // Using the bar's computed side (re-derived here to avoid stale
-        // closures) means the drop always lands where the blue bar was drawn.
-        const insertion = resolveTabInsertion(event, isTabDragData, getDragPointer)
-        if (!insertion) {
-          finishDrag(true)
-          return
-        }
-
-        const overIndex = targetGroup.tabOrder.indexOf(overData.unifiedTabId)
-        const rawInsertIndex = overIndex + (insertion.side === 'right' ? 1 : 0)
-
-        if (activeData.groupId === overData.groupId) {
-          const oldIndex = targetGroup.tabOrder.indexOf(activeData.unifiedTabId)
-          // Why: splicing out the dragged tab before inserting would shift the
-          // intended target slot left by one when moving forward. Adjust the
-          // insertion index to match the post-removal order.
-          const nextIndex = oldIndex < rawInsertIndex ? rawInsertIndex - 1 : rawInsertIndex
-          if (oldIndex !== -1 && oldIndex !== nextIndex) {
-            const nextOrder = targetGroup.tabOrder.filter((id) => id !== activeData.unifiedTabId)
-            nextOrder.splice(nextIndex, 0, activeData.unifiedTabId)
-            reorderUnifiedTabs(overData.groupId, nextOrder)
-            mirrorWebRuntimeTabMove({
-              kind: 'reorder',
-              worktreeId,
-              tabId: activeData.unifiedTabId,
-              targetGroupId: overData.groupId,
-              tabOrder: nextOrder
-            })
-          }
-        } else {
-          const index = overIndex === -1 ? targetGroup.tabOrder.length : rawInsertIndex
-          const moved = dropUnifiedTab(activeData.unifiedTabId, {
-            groupId: overData.groupId,
-            index
-          })
-          if (moved) {
-            shouldRestorePreDragActivation = false
-            mirrorWebRuntimeTabMove({
-              kind: 'move-to-group',
-              worktreeId,
-              tabId: activeData.unifiedTabId,
-              targetGroupId: overData.groupId,
-              index
-            })
-          }
-        }
-
-        finishDrag(
-          shouldRestorePreDragActivation,
-          resolveSourceGroupRestoreOnDrop(
-            activeData,
-            overData.groupId,
-            shouldRestorePreDragActivation
-          )
-        )
-        return
-      }
-
-      if (isPaneDropData(overData)) {
-        if (activeData.groupId !== overData.groupId) {
-          const moved = dropUnifiedTab(activeData.unifiedTabId, {
-            groupId: overData.groupId
-          })
-          if (moved) {
-            shouldRestorePreDragActivation = false
-            mirrorWebRuntimeTabMove({
-              kind: 'move-to-group',
-              worktreeId,
-              tabId: activeData.unifiedTabId,
-              targetGroupId: overData.groupId
-            })
-          }
-        }
-      }
-
-      finishDrag(
-        shouldRestorePreDragActivation,
-        isPaneDropData(overData)
-          ? resolveSourceGroupRestoreOnDrop(
-              activeData,
-              overData.groupId,
-              shouldRestorePreDragActivation
-            )
-          : undefined
-      )
     },
-    [dropUnifiedTab, finishDrag, reorderUnifiedTabs, worktreeId]
+    [dragGeometryRef, dropUnifiedTab, finishDrag, reorderUnifiedTabs, worktreeId]
   )
 
   // Why: dnd-kit fires onDragCancel (not onDragEnd) when the user presses
