+++ src/renderer/src/components/terminal-pane/use-terminal-pane-context-menu.ts
@@ -1,64 +1,27 @@
-/* eslint-disable max-lines -- Why: context-menu actions share pane refs, focus
- * recovery, inherited-cwd split behavior, and agent-fork state in one hook. */
-import { useCallback, useEffect, useRef, useState } from 'react'
-import { toast } from 'sonner'
+import { useCallback, useRef } from 'react'
 import type { ManagedPane, PaneManager } from '@/lib/pane-manager/pane-manager'
 import type { PtyTransport } from './pty-transport'
-import { getConnectionId } from '@/lib/connection-context'
-import { getRuntimeEnvironmentIdForWorktree } from '@/lib/worktree-runtime-owner'
 import type { PaneCwdMap } from './resolve-split-cwd'
 import type { TerminalQuickCommand } from '../../../../shared/terminal-quick-command-types'
 import { isTerminalAgentQuickCommand } from '../../../../shared/terminal-quick-commands'
 import { sendTerminalQuickCommandToPane } from './terminal-quick-command-dispatch'
-import { pasteTerminalText } from './terminal-bracketed-paste'
-import { pasteTerminalClipboard } from './terminal-clipboard-paste'
-import {
-  executeTerminalPastePlan,
-  planTerminalPasteWithYield,
-  type TerminalPasteSource,
-  type TerminalPasteTextOptions
-} from './terminal-paste-coordinator'
-import { formatTerminalPasteExecutionError } from './terminal-paste-errors'
-import { resolveTerminalPasteRuntime } from './terminal-paste-runtime'
-import { getTerminalPasteSshRemotePlatform } from './terminal-paste-ssh-platform'
-import { isTerminalPanePasteTargetCurrent } from './terminal-paste-target-state'
-import { writeTerminalPastePtyInput } from './terminal-pty-paste-writer'
-import { scheduleImagePasteWebglAtlasRecovery } from './terminal-webgl-atlas-recovery'
-import {
-  REQUEST_ACTIVE_TERMINAL_PANE_SPLIT_EVENT,
-  type RequestActiveTerminalPaneSplitDetail
-} from '@/constants/terminal'
-import { makePaneKey } from '../../../../shared/stable-pane-id'
+import type { TerminalPasteSource } from './terminal-paste-coordinator'
 import { runQuickCommandInNewTab } from '@/lib/run-quick-command-in-new-tab'
-import {
-  copyAgentSessionContextFromPane,
-  prepareAgentSessionForkFromPane,
-  type PreparedAgentSessionFork
-} from './terminal-agent-session-fork'
-import { prepareAgentSessionContinuationFromPane } from './terminal-agent-session-continuation'
+import type { PreparedAgentSessionFork } from './terminal-agent-session-fork'
 import type { AgentSessionContinuationRequest } from '@/lib/agent-session-continuation'
-import { recordCreatedTerminalPaneSplit } from './terminal-pane-split-completion'
-import { splitTerminalPaneWithInheritedCwd } from './terminal-pane-split-with-inherited-cwd'
-import { useAppStore } from '@/store'
-import { resolveProtectedMultilinePasteOptionsForPane } from './terminal-agent-paste-bracketing'
-import { resolveTerminalInputHostPlatform } from './terminal-input-host-platform'
-import { translate } from '@/i18n/i18n'
-import { recordTerminalUserInputForLeaf } from './terminal-input-activity'
-import { copyTerminalHandleForPane } from './terminal-handle-copy'
-import { runCopyPaneId, runTerminalCopy } from './terminal-copy-rejection-guards'
-import { copyTerminalSelection } from './terminal-selection-copy'
-
-const CLOSE_ALL_CONTEXT_MENUS_EVENT = 'orca-close-all-context-menus'
-
-export function recordContextMenuCreatedTerminalPaneSplit(
-  createdPane: unknown,
-  args: {
-    source: 'contextual_tour' | 'context_menu'
-    direction: 'vertical' | 'horizontal'
-  }
-): boolean {
-  return recordCreatedTerminalPaneSplit(createdPane, args)
-}
+import { pasteTerminalPaneMenuClipboard } from './terminal-pane-menu-paste'
+import {
+  copyTerminalPaneMenuPaneId,
+  copyTerminalPaneMenuSelection,
+  copyTerminalPaneMenuTerminalId
+} from './terminal-pane-menu-copy-actions'
+import {
+  continueAgentSessionFromMenuPane,
+  copyAgentSessionContextFromMenuPane,
+  forkAgentSessionFromMenuPane
+} from './terminal-pane-menu-agent-session-actions'
+import { useTerminalPaneSplitActions } from './use-terminal-pane-split-actions'
+import { useTerminalContextMenuTrigger } from './use-terminal-context-menu-trigger'
 
 type UseTerminalPaneContextMenuDeps = {
   managerRef: React.RefObject<PaneManager | null>
@@ -131,20 +94,6 @@ export function useTerminalPaneContextMenu({
   rightClickToPaste
 }: UseTerminalPaneContextMenuDeps): TerminalMenuState {
   const contextPaneIdRef = useRef<number | null>(null)
-  const menuOpenedAtRef = useRef(0)
-  const [open, setOpen] = useState(false)
-  const [point, setPoint] = useState({ x: 0, y: 0 })
-
-  useEffect(() => {
-    const closeMenu = (): void => {
-      if (Date.now() - menuOpenedAtRef.current < 100) {
-        return
-      }
-      setOpen(false)
-    }
-    window.addEventListener(CLOSE_ALL_CONTEXT_MENUS_EVENT, closeMenu)
-    return () => window.removeEventListener(CLOSE_ALL_CONTEXT_MENUS_EVENT, closeMenu)
-  }, [])
 
   const resolveMenuPane = useCallback((): ManagedPane | null => {
     const manager = managerRef.current
@@ -159,21 +108,52 @@ export function useTerminalPaneContextMenu({
     return manager.getActivePane() ?? panes[0] ?? null
   }, [managerRef])
 
-  const onCopy = async (): Promise<void> => {
-    const pane = resolveMenuPane()
-    if (!pane) {
-      return
-    }
-    await runTerminalCopy({
-      selection: pane.terminal.getSelection(),
-      writeClipboardText: window.api.ui.writeTerminalClipboardText,
-      // Why: Radix returns focus to the menu trigger (the pane container) on
-      // close, but xterm.js only accepts input when its own helper textarea is
-      // focused. Without this, the user has to click the pane again before
-      // typing works (see #592).
-      focus: () => pane.terminal.focus()
+  const pasteResolvedPane = async (
+    source: Extract<TerminalPasteSource, 'context-menu' | 'right-click'>
+  ): Promise<void> =>
+    pasteTerminalPaneMenuClipboard(
+      {
+        managerRef,
+        paneTransportsRef,
+        tabId,
+        worktreeId,
+        forceBracketedMultilineTextPaste,
+        onPasteError
+      },
+      resolveMenuPane(),
+      source
+    )
+
+  const { open, setOpen, point, menuOpenedAtRef, onContextMenuCapture, onPaneTitleContextMenu } =
+    useTerminalContextMenuTrigger({
+      managerRef,
+      containerRef,
+      contextPaneIdRef,
+      rightClickToPaste,
+      pasteResolvedPane
     })
+
+  const { onSplitRight, onSplitDown } = useTerminalPaneSplitActions({
+    managerRef,
+    paneTransportsRef,
+    paneCwdRef,
+    contextPaneIdRef,
+    tabId,
+    fallbackCwd,
+    resolveMenuPane
+  })
+
+  const agentSessionContext = {
+    paneCwdRef,
+    tabId,
+    worktreeId,
+    groupId,
+    fallbackCwd,
+    onAgentSessionForkReady,
+    onAgentSessionContinuationReady
   }
+
+  const onCopy = async (): Promise<void> => copyTerminalPaneMenuSelection(resolveMenuPane())
 
   const onSelectAll = (): void => {
     const pane = resolveMenuPane()
@@ -183,230 +163,13 @@ export function useTerminalPaneContextMenu({
     }
   }
 
-  const onCopyPaneId = async (): Promise<void> => {
-    const pane = resolveMenuPane()
-    if (!pane) {
-      return
-    }
-    await runCopyPaneId({
-      // Why: orchestration targets use ORCA_PANE_KEY, which survives renderer
-      // remounts; the numeric PaneManager id is only a local runtime handle.
-      paneKey: makePaneKey(tabId, pane.leafId),
-      writeClipboardText: window.api.ui.writeTerminalClipboardText,
-      onSuccess: () =>
-        toast.success(
-          translate(
-            'auto.components.terminal.pane.use.terminal.pane.context.menu.a29b9faa01',
-            'Pane ID copied'
-          )
-        ),
-      // Why: claiming success after a failed write is the exact silent lie this
-      // fallback exists to remove, so report it the way Copy Terminal ID does.
-      onError: () =>
-        toast.error(
-          translate(
-            'auto.components.terminal.pane.use.terminal.pane.context.menu.pane.id.copy.failed',
-            'Unable to copy pane ID'
-          )
-        ),
-      focus: () => pane.terminal.focus()
-    })
-  }
+  const onCopyPaneId = async (): Promise<void> =>
+    copyTerminalPaneMenuPaneId(resolveMenuPane(), tabId)
 
-  const getShortcutPlatform = (): NodeJS.Platform => {
-    if (navigator.userAgent.includes('Mac')) {
-      return 'darwin'
-    }
-    return navigator.userAgent.includes('Windows') ? 'win32' : 'linux'
-  }
-
-  const isPanePasteTargetMounted = (
-    pane: ManagedPane,
-    transport: PtyTransport | undefined,
-    ptyId: string | null
-  ): boolean => {
-    return isTerminalPanePasteTargetCurrent({
-      manager: managerRef.current,
-      paneTransports: paneTransportsRef.current,
-      paneId: pane.id,
-      leafId: pane.leafId,
-      transport,
-      ptyId
-    })
-  }
-
-  const executeMenuPasteText = async (
-    pane: ManagedPane,
-    source: TerminalPasteSource,
-    text: string,
-    options?: TerminalPasteTextOptions
-  ): Promise<boolean> => {
-    const connectionId = getConnectionId(worktreeId) ?? null
-    const transport = paneTransportsRef.current.get(pane.id)
-    const ptyId = transport?.getPtyId() ?? null
-    const shortcutPlatform = getShortcutPlatform()
-    const plan = await planTerminalPasteWithYield({
-      text,
-      source,
-      target: {
-        kind: 'terminal',
-        paneId: pane.id,
-        leafId: pane.leafId,
-        ptyId,
-        runtime: resolveTerminalPasteRuntime({
-          platform: shortcutPlatform,
-          ptyId,
-          connectionId,
-          remotePlatform: getTerminalPasteSshRemotePlatform(connectionId),
-          transport,
-          isWindowsConpty: forceBracketedMultilineTextPaste
-        })
-      },
-      forceBracketedPaste: options?.forceBracketedPaste,
-      forceBracketedPasteForMultiline: options?.forceBracketedPasteForMultiline,
-      windowsInputRecordNewline: options?.windowsInputRecordNewline,
-      terminalBracketedPasteMode: pane.terminal.modes.bracketedPasteMode
-    })
-    const execution = await executeTerminalPastePlan(plan, {
-      pasteText: (pasteText, pasteOptions) =>
-        pasteTerminalText(pane.terminal, pasteText, pasteOptions),
-      writePty: (data) => writeTerminalPastePtyInput(transport, data),
-      isTargetCurrent: () => isPanePasteTargetMounted(pane, transport, ptyId),
-      canContinue: () => isPanePasteTargetMounted(pane, transport, ptyId)
-    })
-    if (execution.status !== 'pasted') {
-      onPasteError(formatTerminalPasteExecutionError(execution.reason))
-      return false
-    }
-    if (text) {
-      recordTerminalUserInputForLeaf(tabId, pane.leafId)
-    }
-    if (options?.recoverImagePasteWebglAtlas) {
-      scheduleImagePasteWebglAtlasRecovery()
-    }
-    return true
-  }
-
-  const onCopyTerminalId = async (): Promise<void> => {
-    const pane = resolveMenuPane()
-    if (!pane) {
-      return
-    }
-    try {
-      await copyTerminalHandleForPane({
-        tabId,
-        leafId: pane.leafId,
-        callRuntime: window.api.runtime.call,
-        writeClipboardText: window.api.ui.writeTerminalClipboardText
-      })
-      toast.success(
-        translate(
-          'auto.components.terminal.pane.use.terminal.pane.context.menu.terminal.id.copied',
-          'Terminal ID copied'
-        )
-      )
-    } catch {
-      toast.error(
-        translate(
-          'auto.components.terminal.pane.use.terminal.pane.context.menu.terminal.id.copy.failed',
-          'Unable to copy terminal ID'
-        )
-      )
-    } finally {
-      pane.terminal.focus()
-    }
-  }
-
-  const pasteResolvedPane = async (
-    source: Extract<TerminalPasteSource, 'context-menu' | 'right-click'>
-  ): Promise<void> => {
-    const pane = resolveMenuPane()
-    if (!pane) {
-      return
-    }
-    const connectionId = getConnectionId(worktreeId) ?? null
-    const state = useAppStore.getState()
-    const runtimeEnvironmentId = getRuntimeEnvironmentIdForWorktree(state, worktreeId)
-    const transport = paneTransportsRef.current.get(pane.id) ?? null
-    const result = await pasteTerminalClipboard({
-      readClipboardText: window.api.ui.readClipboardText,
-      saveClipboardImageAsTempFile: window.api.ui.saveClipboardImageAsTempFile,
-      connectionId,
-      runtimeEnvironmentId,
-      protectedMultilineTextPasteOptions: resolveProtectedMultilinePasteOptionsForPane({
-        isWindowsClient: forceBracketedMultilineTextPaste,
-        hostPlatform: resolveTerminalInputHostPlatform({
-          clientPlatform: getShortcutPlatform(),
-          state,
-          worktreeId,
-          transport
-        }),
-        agentStatusByPaneKey: state.agentStatusByPaneKey,
-        paneForegroundAgentByPaneKey: state.paneForegroundAgentByPaneKey,
-        tabId,
-        leafId: pane.leafId
-      }),
-      pasteText: (text, options) => executeMenuPasteText(pane, source, text, options),
-      onTextPasteError: () =>
-        onPasteError('Paste failed: clipboard text is too large for a safe terminal paste.'),
-      onImagePasteError: (error) => {
-        const detail = error instanceof Error ? error.message : String(error)
-        onPasteError(`Image paste failed: ${detail}`)
-      }
-    })
-    if (result.status !== 'pasted') {
-      return
-    }
-    // Why: Radix returns focus to the menu trigger (the pane container) on
-    // close. Refocus only after a completed paste so rejected async targets
-    // do not steal focus from the user's new control.
-    pane.terminal.focus()
-  }
+  const onCopyTerminalId = async (): Promise<void> =>
+    copyTerminalPaneMenuTerminalId(resolveMenuPane(), tabId)
 
   const onPaste = async (): Promise<void> => pasteResolvedPane('context-menu')
-
-  const splitWithInheritedCwd = useCallback(
-    (
-      direction: 'vertical' | 'horizontal',
-      source: 'contextual_tour' | 'context_menu' = 'context_menu'
-    ): void => {
-      const pane = resolveMenuPane()
-      const manager = managerRef.current
-      if (!pane || !manager) {
-        return
-      }
-      splitTerminalPaneWithInheritedCwd({
-        manager,
-        getManager: () => managerRef.current,
-        paneTransports: paneTransportsRef.current,
-        paneCwdMap: paneCwdRef.current,
-        fallbackCwd,
-        pane,
-        direction,
-        source
-      })
-    },
-    [fallbackCwd, managerRef, paneCwdRef, paneTransportsRef, resolveMenuPane]
-  )
-
-  const onSplitRight = (): void => splitWithInheritedCwd('vertical')
-  const onSplitDown = (): void => splitWithInheritedCwd('horizontal')
-
-  useEffect(() => {
-    const onRequestSplit = (event: Event): void => {
-      const detail = (event as CustomEvent<RequestActiveTerminalPaneSplitDetail>).detail
-      if (detail?.tabId && detail.tabId !== tabId) {
-        return
-      }
-      contextPaneIdRef.current = null
-      splitWithInheritedCwd(detail?.direction ?? 'vertical', getRequestedSplitTelemetrySource())
-    }
-    window.addEventListener(REQUEST_ACTIVE_TERMINAL_PANE_SPLIT_EVENT, onRequestSplit)
-    return () =>
-      window.removeEventListener(REQUEST_ACTIVE_TERMINAL_PANE_SPLIT_EVENT, onRequestSplit)
-    // splitWithInheritedCwd closes over live refs; re-registering keeps the
-    // tour action aligned with the current focused pane and fallback cwd.
-  }, [tabId, splitWithInheritedCwd])
 
   const onEqualizePaneSizes = (): void => {
     const pane = resolveMenuPane()
@@ -432,46 +195,14 @@ export function useTerminalPaneContextMenu({
     }
   }
 
-  const onForkAgentSession = async (): Promise<void> => {
-    const pane = resolveMenuPane()
-    if (!pane) {
-      return
-    }
-    const fork = prepareAgentSessionForkFromPane({ pane, tabId, worktreeId, groupId })
-    if (fork) {
-      onAgentSessionForkReady(fork)
-    }
-  }
+  const onForkAgentSession = async (): Promise<void> =>
+    forkAgentSessionFromMenuPane(agentSessionContext, resolveMenuPane())
 
-  const onContinueAgentSessionInNewSession = (): void => {
-    const pane = resolveMenuPane()
-    if (!pane) {
-      return
-    }
-    const initialCwd = paneCwdRef.current.get(pane.id)?.cwd || fallbackCwd
-    const request = prepareAgentSessionContinuationFromPane({
-      pane,
-      tabId,
-      worktreeId,
-      groupId,
-      workspacePath: fallbackCwd,
-      initialCwd
-    })
-    if (request) {
-      onAgentSessionContinuationReady(request)
-    }
-  }
+  const onContinueAgentSessionInNewSession = (): void =>
+    continueAgentSessionFromMenuPane(agentSessionContext, resolveMenuPane())
 
-  // Why: the captured session transcript is often wanted on its own — to paste
-  // into another tool — so copy the bounded transcript directly, without the
-  // fork prompt's framing or the fork dialog detour (issue #5020).
-  const onCopyAgentSessionContext = async (): Promise<void> => {
-    const pane = resolveMenuPane()
-    if (!pane) {
-      return
-    }
-    await copyAgentSessionContextFromPane(pane)
-  }
+  const onCopyAgentSessionContext = async (): Promise<void> =>
+    copyAgentSessionContextFromMenuPane(resolveMenuPane())
 
   const onQuickCommand = (command: TerminalQuickCommand, historyId: string): void => {
     if (isTerminalAgentQuickCommand(command)) {
@@ -524,77 +255,6 @@ export function useTerminalPaneContextMenu({
     }
   }
 
-  const openContextMenu = (
-    event: React.MouseEvent<HTMLElement>,
-    clickedPaneId: number | null,
-    boundsElement: HTMLElement
-  ): void => {
-    event.preventDefault()
-    window.dispatchEvent(new Event(CLOSE_ALL_CONTEXT_MENUS_EVENT))
-    const manager = managerRef.current
-    if (!manager) {
-      contextPaneIdRef.current = null
-      return
-    }
-    const clickedPane =
-      clickedPaneId !== null
-        ? (manager.getPanes().find((pane) => pane.id === clickedPaneId) ?? null)
-        : null
-    contextPaneIdRef.current = clickedPane?.id ?? null
-
-    // Why: when users opt into terminal-style right-click, a selection copies
-    // and no selection pastes. Ctrl+right-click keeps the app menu reachable.
-    if (rightClickToPaste && !event.ctrlKey) {
-      event.stopPropagation()
-      if (!clickedPane) {
-        return
-      }
-      if (clickedPane.terminal.getSelection()) {
-        void copyTerminalSelection({
-          terminal: clickedPane.terminal,
-          writeClipboardText: window.api.ui.writeTerminalClipboardText,
-          clearSelectionOnSuccess: true
-        }).catch(() => {
-          /* ignore clipboard write failures */
-        })
-      } else {
-        void pasteResolvedPane('right-click')
-      }
-      return
-    }
-
-    menuOpenedAtRef.current = Date.now()
-    const bounds = boundsElement.getBoundingClientRect()
-    setPoint({ x: event.clientX - bounds.left, y: event.clientY - bounds.top })
-    setOpen(true)
-  }
-
-  const onContextMenuCapture = (event: React.MouseEvent<HTMLDivElement>): void => {
-    const manager = managerRef.current
-    if (!manager) {
-      event.preventDefault()
-      contextPaneIdRef.current = null
-      return
-    }
-    const target = event.target
-    if (!(target instanceof Node)) {
-      event.preventDefault()
-      contextPaneIdRef.current = null
-      return
-    }
-    const clickedPane = manager.getPanes().find((pane) => pane.container.contains(target)) ?? null
-    openContextMenu(event, clickedPane?.id ?? null, event.currentTarget)
-  }
-
-  const onPaneTitleContextMenu = (event: React.MouseEvent<HTMLElement>, paneId: number): void => {
-    const boundsElement = containerRef.current
-    if (!boundsElement) {
-      event.preventDefault()
-      return
-    }
-    openContextMenu(event, paneId, boundsElement)
-  }
-
   // Why: PaneManager.getPanes() allocates public pane wrappers. Closed menus
   // do not need pane counts or target identity, so avoid that work on every
   // render across hundreds of mounted terminal tabs.
@@ -629,10 +289,4 @@ export function useTerminalPaneContextMenu({
     onClearPaneTitle: handleClearPaneTitle,
     runForPane
   }
-}
-
-function getRequestedSplitTelemetrySource(): 'contextual_tour' | 'context_menu' {
-  return useAppStore.getState().activeContextualTourId === 'workspace-agent-sessions'
-    ? 'contextual_tour'
-    : 'context_menu'
 }
