+++ src/renderer/src/lib/pane-manager/pane-manager.ts
@@ -1,4 +1,3 @@
-/* eslint-disable max-lines -- Why: PaneManager keeps live pane lifecycle, drag, rendering, and identity callbacks under one owner. */
 import type {
   PaneManagerOptions,
   PaneStyleOptions,
@@ -11,8 +10,8 @@ import type {
   PaneExternalDropTarget
 } from './pane-manager-types'
 import type { SplitPaneAroundLeafIdsOptions } from './pane-subtree-split'
+import type { PaneManagerHost } from './pane-manager-host'
 import {
-  createDivider,
   applyDividerStyles,
   applyPaneOpacity,
   applyRootBackground,
@@ -20,16 +19,9 @@ import {
 } from './pane-divider'
 import { cancelActivePaneDrag, createDragReorderState, handlePaneDrop } from './pane-drag-reorder'
 import { beginPaneDragFromPointerDown } from './pane-drag-pointer'
-import { createPaneDOM, openTerminal, setLigaturesEnabled, disposePane } from './pane-lifecycle'
-import { shouldFollowMouseFocus } from './focus-follows-mouse'
-import { getTerminalWebglAutoDecision } from './terminal-webgl-auto-policy'
-import {
-  equalizePaneSplitSizes,
-  safeFit,
-  fitAllPanesInternal,
-  refitPanesUnder
-} from './pane-tree-ops'
-import { toPublicPane } from './pane-public-view'
+import { setLigaturesEnabled, disposePane } from './pane-lifecycle'
+import { fitAllPanesInternal } from './pane-tree-ops'
+import { collectPublicPanes, toPublicPane } from './pane-public-view'
 import { applyTerminalGpuAcceleration } from './pane-terminal-gpu-acceleration'
 import { rebuildAttachedWebgl } from './pane-webgl-reattach'
 import {
@@ -43,16 +35,28 @@ import type { TerminalLeafId } from '../../../../shared/stable-pane-id'
 import { registerLivePaneManager, unregisterLivePaneManager } from './pane-manager-registry'
 import { releaseHiddenWebglRetention } from './terminal-webgl-hidden-retention'
 import { schedulePaneRevealPresent, schedulePaneRevealRepaint } from './pane-reveal-repaint'
-import { fitRevealedPane } from './pane-reveal-fit'
 import { PaneIdentityRegistry } from './pane-identity-registry'
+import { PaneReparentFrameTracker } from './pane-reparent-frame-tracker'
 import {
-  closeManagedPane,
-  detachManagedPaneForExternalMove,
-  retireManagedPanePreservingPty,
-  splitManagedPane
-} from './pane-split-close'
+  closePaneOnManager,
+  detachPaneForExternalMoveOnManager,
+  retirePanePreservingPtyOnManager,
+  splitPaneAroundLeafIdsOnManager,
+  splitPaneOnManager
+} from './pane-manager-tree-mutations'
+import {
+  createInitialManagedPane,
+  createManagedPaneInternal,
+  publishManagedPaneCreated
+} from './pane-manager-pane-creation'
+import { createManagedPaneDivider, createPaneDragCallbacks } from './pane-manager-drag-wiring'
+import {
+  equalizeManagedPaneSizes,
+  fitRevealedPanes,
+  refreshAllPaneTerminals
+} from './pane-manager-layout-sweeps'
+import { collectPaneRenderingDiagnostics } from './pane-rendering-diagnostics'
 import { FIRST_PANE_ID } from '../../../../shared/pane-key'
-import { splitPaneAroundMountedSubtree } from './pane-subtree-split'
 
 export type {
   PaneManagerOptions,
@@ -75,40 +79,51 @@ export class PaneManager {
   private renderingSuspended: boolean
   private atlasRecoveryVisible: boolean
   private identities = new PaneIdentityRegistry()
-  private pendingPaneReparentFrameIds = new Set<number>()
+  private reparentFrames = new PaneReparentFrameTracker(() => this.destroyed)
 
   // Drag-to-reorder state
   private dragState = createDragReorderState()
+
+  private host: PaneManagerHost
 
   constructor(root: HTMLElement, options: PaneManagerOptions) {
     this.root = root
     this.options = options
     this.renderingSuspended = options.initialRenderingSuspended === true
     this.atlasRecoveryVisible = !this.renderingSuspended
+    this.host = {
+      panes: this.panes,
+      root: this.root,
+      identities: this.identities,
+      dragState: this.dragState,
+      options: this.options,
+      getActivePaneId: () => this.activePaneId,
+      getStyleOptions: () => this.styleOptions,
+      isDestroyed: () => this.destroyed,
+      isRenderingSuspended: () => this.renderingSuspended,
+      allocatePaneId: () => this.nextPaneId++,
+      createPaneInternal: (leafIdHint) => createManagedPaneInternal(this.host, leafIdHint),
+      createDivider: (isVertical) => createManagedPaneDivider(this.host, isVertical),
+      publishPaneCreated: (pane, spawnHints) =>
+        publishManagedPaneCreated(this.host, pane, spawnHints),
+      getDragCallbacks: () => createPaneDragCallbacks(this.host),
+      setActivePane: (paneId, opts) => {
+        this.setActivePane(paneId, opts)
+      },
+      setActivePaneId: (paneId) => {
+        this.activePaneId = paneId
+      },
+      requestPaneReparentFrame: (callback) => {
+        this.reparentFrames.request(callback)
+      }
+    }
     // Why: atlas recovery must reach every live manager — see
     // resetAllTerminalWebglAtlases for the shared-atlas rationale.
     registerLivePaneManager(this)
   }
 
   createInitialPane(opts?: { focus?: boolean; leafId?: string }): ManagedPane {
-    const pane = this.createPaneInternal(opts?.leafId)
-    Object.assign(pane.container.style, {
-      width: '100%',
-      height: '100%',
-      position: 'relative',
-      overflow: 'hidden'
-    })
-    this.root.appendChild(pane.container)
-    openTerminal(pane)
-    this.activePaneId = pane.id
-    applyPaneOpacity(this.panes.values(), this.activePaneId, this.styleOptions)
-
-    if (opts?.focus !== false) {
-      pane.terminal.focus()
-    }
-
-    this.publishPaneCreated(pane)
-    return toPublicPane(pane)
+    return createInitialManagedPane(this.host, opts)
   }
 
   splitPane(
@@ -116,23 +131,7 @@ export class PaneManager {
     direction: 'vertical' | 'horizontal',
     opts?: { ratio?: number; cwd?: string; leafId?: string; ptyId?: string }
   ): ManagedPane | null {
-    return splitManagedPane({
-      paneId,
-      direction,
-      opts,
-      panes: this.panes,
-      root: this.root,
-      styleOptions: this.styleOptions,
-      managerOptions: this.options,
-      createPaneInternal: (leafIdHint) => this.createPaneInternal(leafIdHint),
-      createDivider: (isVertical) => this.createDividerWrapped(isVertical),
-      publishPaneCreated: (pane, spawnHints) => this.publishPaneCreated(pane, spawnHints),
-      getDragCallbacks: () => this.getDragCallbacks(),
-      setActivePaneId: (id) => {
-        this.activePaneId = id
-      },
-      isDestroyed: () => this.destroyed
-    })
+    return splitPaneOnManager(this.host, paneId, direction, opts)
   }
 
   splitPaneAroundLeafIds(
@@ -141,84 +140,29 @@ export class PaneManager {
     direction: 'vertical' | 'horizontal',
     opts?: SplitPaneAroundLeafIdsOptions
   ): ManagedPane | null {
-    return splitPaneAroundMountedSubtree({
+    return splitPaneAroundLeafIdsOnManager(
+      this.host,
       sourceLeafIds,
       fallbackPaneId,
       direction,
-      opts,
-      panes: this.panes,
-      root: this.root,
-      styleOptions: this.styleOptions,
-      managerOptions: this.options,
-      getNumericIdForLeaf: (leafId) => this.identities.getNumericIdForLeaf(leafId),
-      createPaneInternal: (leafIdHint) => this.createPaneInternal(leafIdHint),
-      createDivider: (isVertical) => this.createDividerWrapped(isVertical),
-      publishPaneCreated: (pane, spawnHints) => this.publishPaneCreated(pane, spawnHints),
-      getDragCallbacks: () => this.getDragCallbacks(),
-      setActivePaneId: (id) => {
-        this.activePaneId = id
-      },
-      isDestroyed: () => this.destroyed
-    })
+      opts
+    )
   }
 
   closePane(paneId: number): void {
-    closeManagedPane({
-      paneId,
-      activePaneId: this.activePaneId,
-      panes: this.panes,
-      root: this.root,
-      styleOptions: this.styleOptions,
-      managerOptions: this.options,
-      getDragCallbacks: () => this.getDragCallbacks(),
-      releasePaneIdentity: (numericPaneId) => this.identities.release(numericPaneId),
-      setActivePaneId: (id) => {
-        this.activePaneId = id
-      }
-    })
+    closePaneOnManager(this.host, paneId)
   }
 
   detachPaneForExternalMove(paneId: number): boolean {
-    return detachManagedPaneForExternalMove({
-      paneId,
-      activePaneId: this.activePaneId,
-      panes: this.panes,
-      root: this.root,
-      styleOptions: this.styleOptions,
-      managerOptions: this.options,
-      getDragCallbacks: () => this.getDragCallbacks(),
-      releasePaneIdentity: (numericPaneId) => this.identities.release(numericPaneId),
-      setActivePaneId: (id) => {
-        this.activePaneId = id
-      }
-    })
+    return detachPaneForExternalMoveOnManager(this.host, paneId)
   }
 
   retirePanePreservingPty(paneId: number): boolean {
-    return retireManagedPanePreservingPty({
-      paneId,
-      activePaneId: this.activePaneId,
-      panes: this.panes,
-      root: this.root,
-      styleOptions: this.styleOptions,
-      managerOptions: this.options,
-      getDragCallbacks: () => this.getDragCallbacks(),
-      releasePaneIdentity: (numericPaneId) => this.identities.release(numericPaneId),
-      setActivePaneId: (id) => {
-        this.activePaneId = id
-      }
-    })
+    return retirePanePreservingPtyOnManager(this.host, paneId)
   }
 
   getPanes(limit = Number.POSITIVE_INFINITY): ManagedPane[] {
-    const panes: ManagedPane[] = []
-    for (const pane of this.panes.values()) {
-      if (panes.length >= limit) {
-        break
-      }
-      panes.push(toPublicPane(pane))
-    }
-    return panes
+    return collectPublicPanes(this.panes, limit)
   }
 
   /** Why separate from getPanes: the census runs on the crash path, where
@@ -231,39 +175,16 @@ export class PaneManager {
     fitAllPanesInternal(this.panes)
   }
 
-  // Why: a raw synchronous fit on reveal can apply a transient DOM<->WebGL
-  // cell-metric grid and reflow-garble diff-painting inline TUIs; see fitRevealedPane.
   fitAllRevealedPanes(): void {
-    for (const pane of this.panes.values()) {
-      fitRevealedPane(pane)
-    }
+    fitRevealedPanes(this.panes)
   }
 
   refreshAllPanes(): void {
-    for (const pane of this.panes.values()) {
-      try {
-        if (pane.terminal.rows > 0) {
-          pane.terminal.refresh(0, pane.terminal.rows - 1)
-        }
-      } catch {
-        // Why: restore-all repaint is best-effort while panes are mounting or tearing down.
-      }
-    }
+    refreshAllPaneTerminals(this.panes)
   }
 
   equalizePaneSizes(): void {
-    if (this.panes.size < 2) {
-      return
-    }
-
-    const changed = equalizePaneSplitSizes(
-      this.root.firstElementChild instanceof HTMLElement ? this.root.firstElementChild : null
-    )
-    if (!changed) {
-      return
-    }
-
-    this.options.onLayoutChanged?.()
+    equalizeManagedPaneSizes(this.panes, this.root, this.options.onLayoutChanged)
   }
 
   getActivePane(): ManagedPane | null {
@@ -275,17 +196,7 @@ export class PaneManager {
   }
 
   getRenderingDiagnostics(): PaneRenderingDiagnostics[] {
-    return Array.from(this.panes.values()).map((pane) => ({
-      paneId: pane.id,
-      terminalGpuAcceleration: pane.terminalGpuAcceleration,
-      gpuRenderingEnabled: pane.gpuRenderingEnabled,
-      webglAttachmentDeferred: pane.webglAttachmentDeferred,
-      webglDisabledAfterContextLoss: pane.webglDisabledAfterContextLoss,
-      webglAttachFailedSinceRecovery: pane.webglAttachFailedSinceRecovery === true,
-      hasComplexScriptOutput: pane.hasComplexScriptOutput,
-      terminalWebglAutoDecision: getTerminalWebglAutoDecision(),
-      hasWebgl: Boolean(pane.webglAddon)
-    }))
+    return collectPaneRenderingDiagnostics(this.panes)
   }
 
   hasWebglRenderer(paneId: number): boolean {
@@ -404,11 +315,17 @@ export class PaneManager {
   }
 
   movePane(sourcePaneId: number, targetPaneId: number, zone: DropZone): void {
-    handlePaneDrop(sourcePaneId, targetPaneId, zone, this.dragState, this.getDragCallbacks())
+    handlePaneDrop(sourcePaneId, targetPaneId, zone, this.dragState, this.host.getDragCallbacks())
   }
 
   beginPaneDragFromPointerDown(paneId: number, handle: HTMLElement, event: PointerEvent): void {
-    beginPaneDragFromPointerDown(handle, paneId, this.dragState, this.getDragCallbacks(), event)
+    beginPaneDragFromPointerDown(
+      handle,
+      paneId,
+      this.dragState,
+      this.host.getDragCallbacks(),
+      event
+    )
   }
 
   destroy(): void {
@@ -416,7 +333,7 @@ export class PaneManager {
     unregisterLivePaneManager(this)
     releaseHiddenWebglRetention(this)
     cancelActivePaneDrag(this.dragState)
-    this.cancelPendingPaneReparentFrames()
+    this.reparentFrames.cancelPending()
     for (const pane of this.panes.values()) {
       disposePane(pane, this.panes)
     }
@@ -424,109 +341,5 @@ export class PaneManager {
     disposeDividersIn(this.root)
     this.root.innerHTML = ''
     this.activePaneId = null
-  }
-
-  private createPaneInternal(leafIdHint?: string): ManagedPaneInternal {
-    const id = this.nextPaneId++
-    const leafId = this.identities.claimLeafId(leafIdHint)
-    const pane = createPaneDOM(
-      id,
-      leafId,
-      this.options,
-      this.dragState,
-      this.getDragCallbacks(),
-      // Why: always re-focus even if already active — after splits the
-      // browser's real textarea focus can lag the manager's activePaneId.
-      (paneId, options) => {
-        if (!this.destroyed) {
-          this.setActivePane(paneId, { focus: options?.focusTerminal !== false })
-        }
-      },
-      (paneId, event) => {
-        this.handlePaneMouseEnter(paneId, event)
-      }
-    )
-    pane.webglAttachmentDeferred = this.renderingSuspended
-    this.panes.set(id, pane)
-    this.identities.register(id, leafId)
-    return pane
-  }
-
-  private publishPaneCreated(
-    pane: ManagedPaneInternal,
-    spawnHints?: Parameters<NonNullable<PaneManagerOptions['onPaneCreated']>>[1]
-  ): void {
-    // Why: onPaneCreated wires PTY/status identity synchronously. After this
-    // point, replacing the leaf id would fork ORCA_PANE_KEY from layout state.
-    this.identities.markPublished(pane.id)
-    void this.options.onPaneCreated?.(toPublicPane(pane), spawnHints)
-  }
-
-  private handlePaneMouseEnter(paneId: number, event: MouseEvent): void {
-    if (
-      shouldFollowMouseFocus({
-        featureEnabled: this.styleOptions.focusFollowsMouse ?? false,
-        activePaneId: this.activePaneId,
-        hoveredPaneId: paneId,
-        mouseButtons: event.buttons,
-        windowHasFocus: document.hasFocus(),
-        managerDestroyed: this.destroyed
-      })
-    ) {
-      this.setActivePane(paneId, { focus: true })
-    }
-  }
-
-  private createDividerWrapped(isVertical: boolean): HTMLElement {
-    return createDivider(isVertical, this.styleOptions, {
-      refitPanesUnder: (el) => refitPanesUnder(el, this.panes),
-      onLayoutChanged: this.options.onLayoutChanged,
-      onDragActiveChange: this.options.onPaneDragActiveChange
-    })
-  }
-
-  private getDragCallbacks() {
-    return {
-      getPanes: () => this.panes,
-      getRoot: () => this.root,
-      getStyleOptions: () => this.styleOptions,
-      isDestroyed: () => this.destroyed,
-      safeFit,
-      applyPaneOpacity: () =>
-        applyPaneOpacity(this.panes.values(), this.activePaneId, this.styleOptions),
-      applyDividerStyles: () => applyDividerStyles(this.root, this.styleOptions),
-      refitPanesUnder: (el: HTMLElement) => refitPanesUnder(el, this.panes),
-      requestPaneReparentFrame: (callback: FrameRequestCallback) => {
-        this.requestPaneReparentFrame(callback)
-      },
-      onLayoutChanged: this.options.onLayoutChanged,
-      onDragActiveChange: this.options.onPaneDragActiveChange,
-      resolveExternalDropTarget: this.options.resolveExternalPaneDropTarget,
-      onExternalPaneDrop: this.options.onExternalPaneDrop
-    }
-  }
-
-  private requestPaneReparentFrame(callback: FrameRequestCallback): void {
-    let completed = false
-    let frameId: number | undefined
-    frameId = requestAnimationFrame((timestamp) => {
-      completed = true
-      if (frameId !== undefined) {
-        this.pendingPaneReparentFrameIds.delete(frameId)
-      }
-      if (!this.destroyed) {
-        callback(timestamp)
-      }
-    })
-    if (!completed) {
-      this.pendingPaneReparentFrameIds.add(frameId)
-    }
-  }
-
-  private cancelPendingPaneReparentFrames(): void {
-    for (const frameId of this.pendingPaneReparentFrameIds) {
-      cancelAnimationFrame(frameId)
-    }
-    this.pendingPaneReparentFrameIds.clear()
   }
 }
