+++ src/renderer/src/lib/worktree-activation-issue-command.test.ts
@@ -1,5 +1,5 @@
 import { describe, expect, it } from 'vitest'
-import { ensureWorktreeHasInitialTerminal } from './worktree-activation'
+import { ensureWorktreeHasInitialTerminal } from './worktree-initial-terminal-seeding'
 import {
   createMockStore,
   registerWorktreeActivationReset,
