+++ src/renderer/src/lib/worktree-creation-flow.ts
@@ -1,11 +1,8 @@
 import { toast } from 'sonner'
 import { useAppStore } from '@/store'
 import { TUI_AGENT_CONFIG } from '../../../shared/tui-agent-config'
-import {
-  activateAndRevealWorktree,
-  ensureWorktreeHasInitialTerminal,
-  type ActivateAndRevealResult
-} from '@/lib/worktree-activation'
+import { activateAndRevealWorktree, type ActivateAndRevealResult } from '@/lib/worktree-activation'
+import { ensureWorktreeHasInitialTerminal } from '@/lib/worktree-initial-terminal-seeding'
 import { ensureAgentStartupInTerminal } from '@/lib/new-workspace'
 import { queueWorkspaceActivationTerminalFocus } from '@/lib/workspace-activation-terminal-focus'
 import {
